selectArticlesForExploreDetail
