+++ src/components/explore/explore_section_detail_connected_component.ts
@@ -4,14 +4,14 @@ import { Store } from '../../stores';
 import { ExploreSectionDetailComponent, ExploreSectionDetailProps, ExploreSectionDetailActions } from './explore_section_detail_component';
 import { selectLearn } from '../../selectors/explore';
 import { selectTasksForLearn, selectSavedTasksIdList } from '../../selectors/tasks';
-import { selectArticlesForLearnDetail } from '../../selectors/articles';
+import { selectArticlesForExploreDetail } from '../../selectors/articles/select_articles_for_explore_detail';
 import { RouterProps } from '../../application/routing';
 import { addToSavedList, AddToSavedListAction, Id } from '../../stores/tasks';
 
 const mapStateToProps = (store: Store, ownProps: RouterProps): ExploreSectionDetailProps => ({
     section: selectLearn(store, ownProps),
     tasks: selectTasksForLearn(store, ownProps),
-    articles: selectArticlesForLearnDetail(store, ownProps),
+    articles: selectArticlesForExploreDetail(store, ownProps),
     savedTasksIdList: selectSavedTasksIdList(store),
 });
